+++ src/repositories/mqttRepository.ts
@@ -174,7 +174,8 @@ class MqttRepository {
               capabilityStateInstance: commandTopic.capability?.stateInstance ? commandTopic.capability.stateInstance : '',
               propertyType: commandTopic.property?.type ? commandTopic.property.type : '',
               propertyStateInstance: commandTopic.property?.stateInstance ? commandTopic.property.stateInstance : '',
-              messageValueMapping: commandTopic?.messageValueMapping ? commandTopic?.messageValueMapping : {},
+              topicStateKey: commandTopic.topicStateKey || '',
+              messageValueMapping: commandTopic.messageValueMapping || {},
               userName: parsedTopicName.userName,
             };
             return false;
@@ -435,6 +436,7 @@ export type CommandTopicData = ParsedTopicName & {
   capabilityStateInstance: string;
   propertyType: string;
   propertyStateInstance: string;
+  topicStateKey: string;
   messageValueMapping: { [key: string]: any };
 };